requirements.txt
# Existing dependencies
fastapi
uvicorn
jinja2
python-multipart
psycopg2-binary
python-dotenv

# New dependencies for ARIMA forecasting and analytics
numpy>=1.21.0
pandas>=1.3.0
statsmodels>=0.13.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional: for better performance
scikit-learn>=1.0.0
plotly>=5.0.0
statsmodels
psycopg2-binary
aiofiles
bcrypt
# Email functionality (if using mail_config)
fastapi-mail